refactor: Extract standalone classes from Lalr1Parser.hpp. (#96)
