+++ src/log_surgeon/Lalr1Parser.tpp
@@ -7,6 +7,8 @@
 #include <optional>
 
 #include <log_surgeon/Constants.hpp>
+#include <log_surgeon/ItemSet.hpp>
+#include <log_surgeon/NonTerminal.hpp>
 
 namespace log_surgeon {
 namespace {
