+++ src/log_surgeon/ParserAst.cpp
@@ -1,0 +1,5 @@
+#include <log_surgeon/ParserAst.hpp>
+
+namespace log_surgeon {
+ParserAST::~ParserAST() = default;
+}  // namespace log_surgeon
